devices: vhost-user: fs: Handle child process exit status in fs frontend
The parent process of a vhost-user-fs device was not checking the exit
status of the forked child process so it always exited 0 even when the
fs backend worker killed by a signal.

This change adds proper handling of the child's exit status. This ensures
that failures in the child process are properly propagated as exit code.

BUG=b:435603289
TEST=Kill fs device worker. Then, fs device process exited with 1.

Change-Id: Ie6de5866087ddf309f5020351803d21dd7d38203
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/6817818
Reviewed-by: Frederick Mayle <fmayle@google.com>
Commit-Queue: Keiichi Watanabe <keiichiw@chromium.org>