usb_util: handle EINVAL from DISCARDURB as non-error
The USBDEVFS_DISCARDURB is not an atomic API, in a sense that the user
space may try to discard an URB that is being completed at the same
time. Treat EINVAL as non-error to allow such racy cases.

Since this change can also hide an actual wrong argument, add a debug
message for crosvm development.

BUG=b:417615888
TEST=presubmit
TEST=tast run arc.UsbAttachToArcvm.vm bruschetta.USBStorage

Change-Id: I4258406d35bf8b5a477e5ca591e74009366923c3
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/6928708
Reviewed-by: Ryuichiro Chiba <chibar@chromium.org>
Commit-Queue: Hidenori Kobayashi <hidenorik@chromium.org>