devices: usb: fix completion race for control transfers
When xHCI receives a command transfer with a DataStage, it currently
executes the transfer at the DataStage and sends back completion event
TRB at the StatusStage. However, it is not guaranteed that the xHCI
hardware actually has finished processing the control at this point.
This can cause the guest kernel to emit a warning message that there's
no TD for an event TRB, due to receiving the event TRB for the
StatusStage before that for the DataStage. Fix that by handling the
DataStage together with the StatusStage to properly wait for the
completion event from the host.

BUG=b:417615888
TEST=presubmit
TEST=tast run arc.UsbAttachToArcvm.vm bruschetta.USBStorage

Change-Id: I77e76e99d26331abbbda11a0bec16a91775f6bdc
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/6928705
Reviewed-by: Ryuichiro Chiba <chibar@chromium.org>
Commit-Queue: Hidenori Kobayashi <hidenorik@chromium.org>
Reviewed-by: Keiichi Watanabe <keiichiw@chromium.org>