rutabaga_gfx: rework rutabaga_gfx <--> crosvm CI integration
This CL reworks CI integration.  In the past, the rutabaga_gfx
build.rs script relied on "../third_party/virglrenderer" +
"../third_party/minigbm" being present.  For gfxstream, a
"gfxstream_stub" feature was used.  Ideally, we would get the
dependencies in the CI/CD process, rather than relying
on in-tree checkouts.  This is exactly what the upstream
rutabaga_gfx CI does:

https://github.com/magma-gpu/rutabaga_gfx/pull/2

Is it completely necessary to test these features in crosvm?
Maybe, maybe not.

The features "virglrenderer" and "gfxstream" are completely
internal to rutabaga_gfx, and those are already build tested in
upstream rutabaga_gfx.  And it's already standard operating procedure
to test on hardware-enabled downstream CI's before updating rutabaga
in crosvm.  Both the Android CI and ChromeOS CI are more capable
than both the upstream rutabaga_gfx CI and the crosvm CI.
For example:

https://chromium-review.googlesource.com/q/topic:%22delete-rutabaga-in-chromeos%22

caught issues in the ChromeOS CI, even though it passed the crosvm
CI.

There is the "minigbm" a.k.a "wl-dmabuf" feature though.  That
actually does leak through the rutabaga interface into crosvm,
most notably:

devices/src/virtio/wl.rs

But even THAT could be use rutabaga_gralloc's system gralloc with
some refactoring.  Every CrOS board enables "wl-dmabuf" except Grunt,
but the issue there is nobody is going to test with a Grunt.  Plus,
one can probably replace virtio-wl with virtio-gpu entirely now.

Either way, with the proviso that we COULD probably optimize/clean-up
further, this just adds aarch64/x86_64 builds of gfxstream, virglrenderer
and minigbm to the dev_container.  The files are copied over to the VMs
from the container when necessary.

The "gfxstream_stub" feature is removed since it was used to accomodate
the crosvm CI anyways.

dev_container version bumper to 0053.

BUG=438557794
TEST=
     tools/presubmit
     tools/dev_container cargo build --features=gfxstream,virgl_renderer,wl-dmabuf
     tools/dev_container cargo build --target=aarch64-unknown-linux-gnu \
     --features=gfxstream,virgl_renderer,wl-dmabuf

Change-Id: Id7fd84a8123fb2b9fa6c255a1102019456eb2607
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/6886676
Reviewed-by: Frederick Mayle <fmayle@google.com>
Commit-Queue: Gurchetan Singh <gurchetansingh@chromium.org>