x86_64: force disabling balloon for pVMs
Supporting memory balloon for protected VMs requires special support in
pKVM and in the guest, to let the protected guest ask pKVM via a
hypercall to give the host access to the guest's pages reclaimed by the
host. This is implemented in pKVM on arm64 but not on x86 yet. So
temporarily force disabling balloon for protected VMs on x86, until such
support is available.

BUG=b:438723514, b:416305513
TEST=run a Linux pVM and check that is has no balloon device in lspci

Change-Id: I92a220c00ddab6d09fc0fa09c2da74efc8c03f82
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/6911146
Reviewed-by: Grzegorz Jaszczyk <jaszczyk@google.com>
Reviewed-by: Dmitry Torokhov <dtor@chromium.org>
Commit-Queue: Dmytro Maluka <dmaluka@chromium.org>