crosvm: remove unnecessary GPU conditionals
There are too many GPU-related conditionals in crosvm.

For example, GpuMode is an enum with no dependency on
libgfxstream_backend.so and libvirglrenderer.so, so
it should be always compilable.  If the user selects
one of those GPU modes and doesn't build rutabaga with
it, rutabaga_gfx will complain loudly.

This simplifies the testing logic in gpu_config.

In addition, some of the tests need to be modified.
For example, it's FINE for the gfxstream to be the
default on Windows, but 2D mode is always the simplest
to get working and should be a valid option.  Note 2D
mode doesn't work right now on upstream crosvm on
Windows, but neither does gfxstream + upstream
crosvm (though it does with gfxstream + downstream crosvm).

BUG=419304200
TEST=CI/CD

Change-Id: I96731f7664222de8ec48f29349adaf7670dc7e84
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/6967408
Reviewed-by: Kaiyi Li <kaiyili@google.com>
Reviewed-by: Frederick Mayle <fmayle@google.com>
Commit-Queue: Gurchetan Singh <gurchetansingh@chromium.org>