devices: gpu: remove more unnecessary conditionals
No need to use a conditional when we can use the rutabaga
component type directly.  An keen observer might comment
that this drops the number of capability sets from 4 to
3 when both gfxstream/virglrenderer are enabled.  Yes,
that is correct.  But nobody really runs them at the same
time and does interop between, so this is fine.

BUG=419304200
TEST=CI/CD

Change-Id: I4e6fe80eb5fca0b132124f662ab3b7b52b58caff
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/6985924
Reviewed-by: Frederick Mayle <fmayle@google.com>
Commit-Queue: Gurchetan Singh <gurchetansingh@chromium.org>