crosvm: windows: cmdline: delete unnecessary conditional
This is a commandline parsing test.

The availability of gfxstream_backend as no effect on this.
Besides, B* enables gfxstream by default on Windows.

"angle" is deleted since it's not in GpuParameters.  Similar
reasons for the other modifications: seems like code that
wasn't tested before.

BUG=444309893
TEST=CI/CD

Change-Id: I131c6bac0ea4a3efae43a77fb36a1810d8886564
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/6985926
Reviewed-by: Frederick Mayle <fmayle@google.com>
Commit-Queue: Gurchetan Singh <gurchetansingh@chromium.org>
Reviewed-by: Kaiyi Li <kaiyili@google.com>