devices: vhost: always set vring_call for each queue
The init routine and the config change routine were inconsistent. On
init, if the MSI-X config masked a vector, then we didn't register an
interrupt with vhost at all. If a vector is masked, we are supposed to
set the pending bit so that it can be sent once unmasked, so this
potentially could have resulted in dropped interrupts.

BUG=b:401030573

Change-Id: Idd5db3d27c22ceaa936f0b1c10c58fac4818724f
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/6997987
Reviewed-by: Keiichi Watanabe <keiichiw@chromium.org>
Commit-Queue: Frederick Mayle <fmayle@google.com>