Don't compare backtraces of errors.
If the VmMemoryResponseError has a backtrace (which it will if built
with a recent version of anyhow and RUST_BACKTRACE=1) then it is not
serialised or deserialised, so the test fails. To fix this, rather than
comparing the Debug format of the error (which includes the backtrace),
just compare the Display format of the error and its causes.

BUG=b:384438979
TEST=`RUST_BACKTRACE=1 cargo test` with anyhow 1.0.100

Change-Id: Ic7d9466957ce08f68b4911f5c2648cb23b98a3c7
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/7062577
Auto-Submit: Andrew Walbran <qwandor@google.com>
Reviewed-by: Frederick Mayle <fmayle@google.com>
Commit-Queue: Frederick Mayle <fmayle@google.com>