devices: virtio: block: VIRTIO_BLK_T_GET_ID: default to empty
If no block id is provided for a disk, respond with an empty ID instead
of an "unsupported" error. At least in Linux, the guest driver treats
"unsupported" as an empty ID anyways, so there is no behavior
difference.

This gets rid of some distracting debug logs + simplifies the code.

Change-Id: I52a90b93e10e96f66f35d61860b5d50dd0440309
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/7089028
Reviewed-by: Dmitry Torokhov <dtor@chromium.org>
Commit-Queue: Frederick Mayle <fmayle@google.com>