crosvm: delete config-file feature
We are unlikely to prioritize finishing this feature and haven't been
enforcing stability of the options in the config as intended.

Deleting has some benefits:

* Less code.
* Deadcode analysis on fields works again (because we don't have both
  Deserialize and Serialize implementations). Turns out at least one
  field was not wired up to the corresponding Config field.
* Don't need to worry about cmdline and config defaults diverging (argh
  and serde choose defaults differently, not everyone was following the
  "must use Option or Vec" advise documented on the top level cmdline
  struct). Note that this only helps with top level fields, complex
  fields using `FromKeyValues` still depend on serde.

To avoid deadcode lint errors, some cmdline options are now build time
conditional. There is a possiblity this breaks someone invocation of
crosvm, but, if so, there was likely a bug because that option had no
effect before.

BUG=460201654

Change-Id: Ia7a1d31f7573120d2d6d289ecb3549237db7a0d7
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/7102435
Commit-Queue: Frederick Mayle <fmayle@google.com>
Reviewed-by: Keiichi Watanabe <keiichiw@chromium.org>