vfio: Refactor VFIO_DEVICE_FEATURE into helper
In anticipation of more calls to the ioctl only differing in `flags` and
to avoid the proliferation of unsafe code, De-dup the raw iotclt call
into a helper, which is safe because any value for flags results in a
safe call (at worst, it results in an error from the kernel).

Note: No functional change intended.

BUG=b:460349550
TEST=tools/presubmit --all

Change-Id: I6bec7ae6d64a4522505c464ca0b756f9742367e7
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/7209064
Auto-Submit: Pierre-Clément Tosi <ptosi@google.com>
Reviewed-by: Frederick Mayle <fmayle@google.com>
Commit-Queue: Frederick Mayle <fmayle@google.com>