Revert "x86_64: force disabling balloon for pVMs"
This reverts commit 3202c663eaa555678bd142320d51c63acce266bc.

While memory balloon for protected VMs is still not supported by pKVM on
x86 yet, it can be disabled by the guest itself (as the guest is aware
that it is running as a pKVM's protected VM anyway), so it is not
necessary to force disabling it in crosvm.

Now that the guest (microdroid) kernel prebuilt in AVF has been updated
to prevent enabling the balloon when running as a pVM, we can remove
disabling it in crosvm.

BUG=b:438723514
TEST=run a Linux pVM and check that is has no balloon device in lspci

Change-Id: Ic4ab825b9f63746d324e7d8dc89a2ed53d933a59
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/7246100
Reviewed-by: Frederick Mayle <fmayle@google.com>
Reviewed-by: Grzegorz Jaszczyk <jaszczyk@google.com>
Commit-Queue: Dmytro Maluka <dmaluka@chromium.org>