devices: Add infrastructure for hypercall handling
Add generic support for implementing hypercall handlers in CrosVM.

Introduce the concept of a hypercall bus, enabling handlers to be
presented as devices on that bus, similar to the existing MMIO bus and
IO bus. For now, to enable reusing the internal device look-up logic,
simply introduce the new BusType::Hypercall and the corresponding method
handle_hypercall(). Ideally, this will eventually be cleaned up e.g.
with a separate struct for each bus and a possible shared trait.

Enable full flexibility regarding the hypervisor and CPU in use with the
following indirection:

1. platform-specific exits are converted into VcpuExit::Hypercall
2. Vcpu::handle_hypercall() is called and can populate an
   architecture-specific HypercallAbi in a hypervisor-specific way
3. Bus::handle_hypercall() is then called on the result in a generic way
4. the bus looks up the device and calls BusDevice::handle_hypercall()
5. the device can handle the call in a hypervisor-agnostic way, only
   having to deal with the ABI, typically through architecture-agnostic
   code

This results in device code that can ideally be re-used across
hypervisors and can typically be built for any architecture, removing
the need for more complexity related to conditional compilation.

BUG=b:460349550
TEST=tools/presubmit

Change-Id: Idf59294597d66a91f922ac5e78dae9095e33caea
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/7253050
Reviewed-by: Frederick Mayle <fmayle@google.com>
Commit-Queue: Frederick Mayle <fmayle@google.com>
Auto-Submit: Pierre-Clément Tosi <ptosi@google.com>